Move network utils code in net directory (#960)
* Move network utils code in net directory

* Repair build

* Repair build on Android
