+++ src/libmeasurement_kit/common/utils.cpp
@@ -4,13 +4,10 @@
 
 #include "../common/utils.hpp"
 
+#include <event2/util.h>
+
 #include <ctype.h>
 #include <math.h>
-
-#include <deque>
-#include <cstring>
-
-#define MEASUREMENT_KIT_SOCKET_INVALID -1
 
 namespace mk {
 
@@ -48,111 +45,6 @@ timeval *timeval_init(timeval *tv, double delta) {
     tv->tv_sec = (time_t)floor(delta);
     tv->tv_usec = (suseconds_t)((delta - floor(delta)) * 1000000);
     return tv;
-}
-
-int storage_init(sockaddr_storage *storage, socklen_t *salen,
-                 const char *family, const char *address, const char *port) {
-    int _family;
-    /* TODO: support also AF_INET, AF_INET6, ... */
-    if (strcmp(family, "PF_INET") == 0) {
-        _family = PF_INET;
-    } else if (strcmp(family, "PF_INET6") == 0) {
-        _family = PF_INET6;
-    } else {
-        warn("utils:storage_init: invalid family");
-        return -1;
-    }
-    return storage_init(storage, salen, _family, address, port);
-}
-
-int storage_init(sockaddr_storage *storage, socklen_t *salen, int _family,
-                 const char *address, const char *port) {
-    const char *errstr;
-    int _port = (int)mkp_strtonum(port, 0, 65535, &errstr);
-    if (errstr != nullptr) {
-        warn("utils:storage_init: invalid port");
-        return -1;
-    }
-    return storage_init(storage, salen, _family, address, _port);
-}
-
-int storage_init(sockaddr_storage *storage, socklen_t *salen, int _family,
-                 const char *address, int _port) {
-    int result;
-
-    debug("utils:storage_init - enter");
-
-    if (_port < 0 || _port > 65535) {
-        warn("utils:storage_init: invalid port");
-        return -1;
-    }
-
-    memset(storage, 0, sizeof(*storage));
-    switch (_family) {
-
-    case PF_INET6: {
-        sockaddr_in6 *sin6 = (sockaddr_in6 *)storage;
-        sin6->sin6_family = AF_INET6;
-        sin6->sin6_port = htons(_port);
-        if (address != nullptr) {
-            result = inet_pton(AF_INET6, address, &sin6->sin6_addr);
-            if (result != 1) {
-                warn("utils:storage_init: invalid addr");
-                return -1;
-            }
-        } else {
-            sin6->sin6_addr = in6addr_any;
-        }
-        *salen = sizeof(sockaddr_in6);
-        break;
-    }
-
-    case PF_INET: {
-        sockaddr_in *sin = (sockaddr_in *)storage;
-        sin->sin_family = AF_INET;
-        sin->sin_port = htons(_port);
-        if (address != nullptr) {
-            result = inet_pton(AF_INET, address, &sin->sin_addr);
-            if (result != 1) {
-                warn("utils:storage_init: invalid addr");
-                return -1;
-            }
-        } else {
-            sin->sin_addr.s_addr = INADDR_ANY;
-        }
-        *salen = sizeof(sockaddr_in);
-        break;
-    }
-
-    default:
-        throw std::runtime_error("invalid case");
-    }
-
-    debug("utils:storage_init - ok");
-    return 0;
-}
-
-evutil_socket_t socket_create(int domain, int type, int protocol) {
-    evutil_socket_t filedesc;
-    int result;
-
-    debug("utils:socket - enter");
-
-    filedesc = socket(domain, type, protocol);
-    if (filedesc == MEASUREMENT_KIT_SOCKET_INVALID) {
-        warn("utils:socket: cannot create socket");
-        return MEASUREMENT_KIT_SOCKET_INVALID;
-    }
-
-    result = evutil_make_socket_nonblocking(filedesc);
-    if (result != 0) {
-        warn("utils:socket: cannot make nonblocking");
-        (void)evutil_closesocket(filedesc);
-        return MEASUREMENT_KIT_SOCKET_INVALID;
-    }
-
-    debug("utils:socket - ok");
-    return filedesc;
 }
 
 // See <http://stackoverflow.com/questions/440133/>
@@ -198,59 +90,6 @@ std::string random_str_uppercase(size_t length) {
             "ABCDEFGHIJKLMNOPQRSTUVWXYZ"  // uppercase
         ;
     return random_within_charset(num_upper, length);
-}
-
-std::string unreverse_ipv6(std::string s) {
-    size_t i = 0, added = 0;
-    std::deque<char> r;
-    for (; i < s.size(); ++i) {
-        if ((i % 2) == 0) {
-            if (!isxdigit(s[i])) {
-                break;
-            }
-            r.push_front(s[i]);
-            if ((++added % 4) == 0 && added <= 28) {
-                r.push_front(':');
-            }
-        } else {
-            if (s[i] != '.') {
-                return "";
-            }
-        }
-    }
-    if (s.substr(i) != "ip6.arpa" && s.substr(i) != "ip6.arpa.") {
-        return "";
-    }
-    return std::string(r.begin(), r.end());
-}
-
-std::string unreverse_ipv4(std::string s) {
-    std::deque<char> r, t;
-    size_t i = 0, seen = 0;
-    unsigned cur = 0;
-    for (; i < s.size(); ++i) {
-        if (s[i] == '.') {
-            if (cur > 255) {
-                return "";
-            }
-            if (seen++ > 0) {
-                r.push_front('.');
-            }
-            r.insert(r.begin(), t.begin(), t.end());
-            t.clear();
-            cur = 0;
-        } else if (isdigit(s[i])) {
-            t.push_back(s[i]);
-            char tmpstr[] = {s[i], '\0'};
-            cur = cur * 10 + atoi(tmpstr);
-        } else {
-            break;
-        }
-    }
-    if (s.substr(i) != "in-addr.arpa" && s.substr(i) != "in-addr.arpa.") {
-        return "";
-    }
-    return std::string(r.begin(), r.end());
 }
 
 void dump_settings(Settings &s, std::string prefix, Var<Logger> logger) {
