+++ src/libmeasurement_kit/net/utils.hpp
@@ -1,0 +1,54 @@
+// Part of measurement-kit <https://measurement-kit.github.io/>.
+// Measurement-kit is free software. See AUTHORS and LICENSE for more
+// information on the copying conditions.
+#ifndef SRC_LIBMEASUREMENT_KIT_NET_UTILS_HPP
+#define SRC_LIBMEASUREMENT_KIT_NET_UTILS_HPP
+
+#include <measurement_kit/net.hpp>
+
+struct sockaddr_storage;
+
+namespace mk {
+namespace net {
+
+int storage_init(
+        sockaddr_storage *storage,
+        socklen_t *length,
+        const char *family,
+        const char *address,
+        const char *port,
+        Var<Logger> logger
+);
+
+int storage_init(
+        sockaddr_storage *storage,
+        socklen_t *length,
+        int family,
+        const char *address,
+        const char *port,
+        Var<Logger> logger
+);
+
+int storage_init(
+        sockaddr_storage *storage,
+        socklen_t *length,
+        int family,
+        const char *address,
+        int port,
+        Var<Logger> logger
+);
+
+socket_t socket_create(
+        int domain,
+        int type,
+        int protocol,
+        Var<Logger> logger
+);
+
+std::string unreverse_ipv6(std::string s);
+
+std::string unreverse_ipv4(std::string s);
+
+} // namespace net
+} // namespace mk
+#endif
